# packages/api-client/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Smartnet.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/invoices": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceSummary"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceSummary"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Invoices"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateInvoiceRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateInvoiceRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateInvoiceRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceCreatedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceCreatedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceCreatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoices/{id}": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoices/tax-rate": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "companyId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceTaxRate"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceTaxRate"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceTaxRate"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoices/credit-status": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "companyId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreditStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditStatus"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInvoiceLineRequest": {
        "required": [
          "discountPercent",
          "quantity",
          "unitPrice"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "itemCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "discountPercent": {
            "type": "number",
            "format": "double"
          },
          "cost": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateInvoiceRequest": {
        "required": [
          "acknowledgeCreditLimit",
          "companyId",
          "customerId",
          "date",
          "documentDiscountPercent",
          "lines",
          "type"
        ],
        "type": "object",
        "properties": {
          "companyId": {
            "type": "integer",
            "format": "int64"
          },
          "customerId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "purchaseOrderNo": {
            "type": "string",
            "nullable": true
          },
          "contactPerson": {
            "type": "string",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateInvoiceLineRequest"
            }
          },
          "documentDiscountPercent": {
            "type": "number",
            "format": "double"
          },
          "acknowledgeCreditLimit": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CreditStatus": {
        "required": [
          "creditLimit",
          "enforced",
          "outstanding"
        ],
        "type": "object",
        "properties": {
          "creditLimit": {
            "type": "number",
            "format": "double"
          },
          "outstanding": {
            "type": "number",
            "format": "double"
          },
          "enforced": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceCreatedResponse": {
        "required": [
          "id",
          "number",
          "outstanding",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "outstanding": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "InvoiceDetail": {
        "required": [
          "date",
          "discountAmount",
          "id",
          "kind",
          "lines",
          "netTotal",
          "number",
          "origin",
          "outstanding",
          "subtotal",
          "taxAmount",
          "taxRatePercentage",
          "total",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "type": {
            "type": "string"
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "kind": {
            "type": "string"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "customerCode": {
            "type": "string",
            "nullable": true
          },
          "purchaseOrderNo": {
            "type": "string",
            "nullable": true
          },
          "contactPerson": {
            "type": "string",
            "nullable": true
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "discountAmount": {
            "type": "number",
            "format": "double"
          },
          "netTotal": {
            "type": "number",
            "format": "double"
          },
          "taxRatePercentage": {
            "type": "number",
            "format": "double"
          },
          "taxAmount": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "outstanding": {
            "type": "number",
            "format": "double"
          },
          "origin": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineDetail"
            }
          }
        },
        "additionalProperties": false
      },
      "InvoiceLineDetail": {
        "required": [
          "discountPercent",
          "gross",
          "net",
          "quantity",
          "unitPrice"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "itemCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "discountPercent": {
            "type": "number",
            "format": "double"
          },
          "gross": {
            "type": "number",
            "format": "double"
          },
          "net": {
            "type": "number",
            "format": "double"
          },
          "cost": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InvoiceSummary": {
        "required": [
          "date",
          "id",
          "number",
          "origin",
          "outstanding",
          "total",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "outstanding": {
            "type": "number",
            "format": "double"
          },
          "origin": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "InvoiceTaxRate": {
        "required": [
          "name",
          "percentage"
        ],
        "type": "object",
        "properties": {
          "taxRateId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "percentage": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}